Add EventHandler for future other DM Events
I have re-add an EventHandler , full debug

Xephi
