+++ src/com/timvisee/DungeonMaze/API/DungeonMazeAPI.java
@@ -8,7 +8,7 @@ import org.bukkit.entity.Player;
 import org.bukkit.plugin.Plugin;
 
 import com.timvisee.DungeonMaze.DungeonMaze;
-import com.timvisee.DungeonMaze.event.generation.DMGenerationChestEvent;
+import com.timvisee.DungeonMaze.event.EventHandler.DMEventHandler;
 import com.timvisee.DungeonMaze.manager.DMWorldManager;
 
 public class DungeonMazeAPI {
@@ -176,7 +176,7 @@ public class DungeonMazeAPI {
 	}
 	
 	
-	public static DMGenerationChestEvent setupDMEventHandler() {
-		return (DMGenerationChestEvent) DMGenerationChestEvent.getServer();
+	public static DMEventHandler setupDMEventHandler() {
+		return (DMEventHandler) DMEventHandler.getServer();
 	}
 }
